+++ jest.setup.js
@@ -1,0 +1,2 @@
+// Add any global setup here
+import '@testing-library/jest-dom';